+++ RELEASES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## [Unreleased]
+
+### Changed
+
+* [msr] Updated for the Intel Software Developer's Manual, September 2023 release
+
+
 ## [v0.9.0] - 2023-04-30
 
 ### Added
@@ -263,6 +270,7 @@
 * Initial public release
 
 
+[Unreleased]: https://github.com/powercap/raplcap/compare/v0.9.0...HEAD
 [v0.9.0]: https://github.com/powercap/raplcap/compare/v0.8.0...v0.9.0
 [v0.8.0]: https://github.com/powercap/raplcap/compare/v0.7.0...v0.8.0
 [v0.7.0]: https://github.com/powercap/raplcap/compare/v0.6.0...v0.7.0